requirements.txt
SpeechRecognition==3.9.0
discord.py>=1.7.3
discord-ext-audiorec>=0.1.6
